+++ src/main/webapp/wise5/components/audioOscillator/audioOscillatorAuthoringComponentModule.es6
@@ -1,0 +1,20 @@
+'use strict';
+
+import AudioOscillatorService from './audioOscillatorService';
+import AudioOscillatorController from './audioOscillatorController';
+import AudioOscillatorAuthoringController from './audioOscillatorAuthoringController';
+
+let audioOscillatorAuthoringComponentModule = angular.module('audioOscillatorAuthoringComponentModule', [
+  'pascalprecht.translate'
+])
+  .service(AudioOscillatorService.name, AudioOscillatorService)
+  .controller(AudioOscillatorController.name, AudioOscillatorController)
+  .controller(AudioOscillatorAuthoringController.name, AudioOscillatorAuthoringController)
+  .config([
+    '$translatePartialLoaderProvider',
+    ($translatePartialLoaderProvider) => {
+      $translatePartialLoaderProvider.addPart('components/audioOscillator/i18n');
+    }
+  ]);
+
+export default audioOscillatorAuthoringComponentModule;
